fix: address review comments - shared formatBytes, listener cleanup, pagination
