+++ packages/shared/src/utils/file-utils.ts
@@ -294,4 +294,15 @@ export class FileUtils {
 
     return normalizedPath.startsWith(normalizedBase);
   }
+
+  /**
+   * Format bytes to human-readable string (e.g., "1.5 MB")
+   */
+  static formatBytes(bytes: number): string {
+    if (bytes === 0) return "0 Bytes";
+    const k = 1024;
+    const sizes = ["Bytes", "KB", "MB", "GB", "TB"];
+    const i = Math.floor(Math.log(bytes) / Math.log(k));
+    return parseFloat((bytes / Math.pow(k, i)).toFixed(2)) + " " + sizes[i];
+  }
 }
